+++ frontend/src/features/opportunity/components/KanbanBoard.tsx
@@ -22,7 +22,7 @@ import FilterListIcon from '@mui/icons-material/FilterList';
 import CheckCircleIcon from '@mui/icons-material/CheckCircle';
 import CancelIcon from '@mui/icons-material/Cancel';
 
-import { OpportunityStage, STAGE_CONFIGS } from '../types/stages';
+import { OpportunityStage, STAGE_CONFIGS, type Opportunity } from '../types';
 import { logger } from '../../../lib/logger';
 import { useErrorHandler } from '../../../components/ErrorBoundary';
 
@@ -40,18 +40,7 @@ const CLOSED_STAGES = [
   OpportunityStage.CLOSED_LOST,
 ];
 
-interface Opportunity {
-  id: string;
-  name: string;
-  stage: OpportunityStage;
-  value?: number;
-  probability?: number;
-  customerName?: string;
-  assignedToName?: string;
-  expectedCloseDate?: string;
-  createdAt: string;
-  updatedAt: string;
-}
+// Type imported from '../types'
 
 // Convert STAGE_CONFIGS array to Record for easier lookup
 const STAGE_CONFIGS_RECORD: Record<string, typeof STAGE_CONFIGS[0]> = {};
